# Conflicts:
#	components/header/index.js
#	components/sidebar/index.js
#	components/table/pagination.js
#	pages/hoc-sinh/index.js
#	pages/nhan-su/nhan-vien/them.js
#	pages/to-chuc/khoi/[id].js
#	pages/to-chuc/khoi/index.js
#	pages/to-chuc/khoi/them.js
#	pages/to-chuc/lop/index.js
#	pages/to-chuc/nien-khoa/[id].js
#	pages/to-chuc/nien-khoa/index.js
#	pages/to-chuc/nien-khoa/them.js
#	pages/to-chuc/truong/[id].js
#	pages/to-chuc/truong/index.js
#	pages/to-chuc/truong/them.js
#	services/organize/schoolYear.js

+++ pages/to-chuc/truong/index.js
@@ -1,66 +1,15 @@
 import {useEffect, useState} from "react";
 import Link from "next/link";
 import {useRouter} from "next/router";
-
 import {PencilIcon, TrashIcon} from "@heroicons/react/outline";
 import Input from "@components/form/input";
-import Layout from "@components/layout";
-import Button from "@components/button";
+import Pagination from "@components/pagination";
 import {schoolService} from "@services";
-import Table from "@components/table";
 
 const SchoolList = () => {
-
   const [schools, setSchools] = useState([])
   const router = useRouter();
-
-  const columns = [
-    {
-      id: 'id',
-      title: 'STT',
-      key: 'id'
-    },
-    {
-      id: 'schoolname',
-      title: 'Tên trường',
-    },
-    {
-      id: 'address',
-      title: 'Địa chỉ',
-    },
-    {
-      id: 'province',
-      title: 'Tỉnh',
-    },
-    {
-      id: 'district',
-      title: 'Quận',
-    },
-    {
-      id: 'ward',
-      title: 'Phường',
-    },
-    {
-      id: 'civilGroup',
-      title: 'Nhóm trường',
-    },
-    {
-      id: 'action',
-      title: 'Thao tác',
-      render: (element) => (
-        <>
-          <Link href={router.pathname + '/' + element._id}>
-            <a><PencilIcon className='h-5 w-5 inline'/></a>
-          </Link>
-          <TrashIcon
-            className='h-5 w-5 inline ml-4 cursor-pointer'
-            onClick={() => handleDelete(element._id)}
-          />
-        </>
-      )
-    }
-  ]
-
+  let skip = 0;
   useEffect(async () => {
     try {
       const {...response} = await schoolService.list()
@@ -71,27 +20,74 @@ const SchoolList = () => {
   }, []);
 
   const handleDelete = async (id) => {
-    try {
-      await schoolService.delete(id)
-      await swal('Xoá thành công');
-      router.reload();
-    } catch (error) {
-      console.log({error})
-    }
+    swal({
+      title: "Bạn chắc chắn muốn xóa?",
+      text: "",
+      icon: "warning",
+      buttons: true,
+      successMode: true,
+    }).then(async (willDelete) => {
+      if (willDelete) {
+        const result = await schoolService.delete(id);
+        if(result){
+          router.reload();
+        }
+        else{
+          swal('Xóa không thành công!!', '', 'error');s
+        }
+      }
+    });
+  
   };
 
   return (
     <>
       <h4>Tổ chức</h4>
-      <Input className='md:w-1/2 lg:w-1/4' placeholder='Tìm kiếm...'/>
-      <Link href={router.pathname + '/' + 'them'}>
-        <a><Button>Thêm mới</Button></a>
-      </Link>
-      <Table columns={columns} rows={schools} widthContainer='w-[1200px]'/>
+      <Input className='md:w-1/2 lg:w-1/4' placeholder='Tìm kiếm...' name="s"/>
+      <div className="mt-8 overflow-x-auto lg:overflow-x-visible">
+        <div className='container-table'>
+          <h4>Danh sách trường</h4>
+          <table className='table'>
+            <thead>
+              <tr>
+                <th className='text-center'>STT</th>
+                <th>Tên trường</th>
+                <th>Địa chỉ</th>
+                <th>Tỉnh/thành</th>
+                <th>Quận/Huyện</th>
+                <th>Phường/Xã</th>
+                <th/>
+              </tr>
+            </thead>
+            <tbody>
+              {schools?.map((school, index) => (
+                <tr key={index}>
+                  <td>{skip + index + 1}</td>
+                  <td>{school.schoolname}</td>
+                  <td>{school.address}</td>
+                  <td>{school.province?.provinceName}</td>
+                  <td>{school.district?.districtName}</td>
+                  <td>{school.ward?.wardName}</td>
+                  <td>
+                     <Link href={`/to-chuc/truong/${school._id}`}>
+                       <a><PencilIcon className='h-5 w-5 inline'/></a>
+                     </Link>
+                     <TrashIcon
+                       className='h-5 w-5 inline ml-4 cursor-pointer'
+                       onClick={() => handleDelete(school._id)}
+                     />
+                  </td>
+              </tr>
+              ))}
+            </tbody>
+          </table>
+          {/* <Pagination data={schools}/> */}
+        </div>
+      </div>
+
     </>
   );
 }
 
 export default SchoolList;
 
-SchoolList.getLayout = (page) => <Layout>{page}</Layout>;
